+++ scripts/macros/2024/classFeatures/bard/jackOfAllTrades.js
@@ -1,0 +1,5 @@
+export let jackOfAllTrades = {
+    name: 'Jack of All Trades',
+    version: '1.1.30',
+    rules: 'modern'
+};